Add the lifetime extension system
This enables garbage collection across language boundaries, when the C
part is compiled with pointer trap support.

At this point the support is still incomplete: the Python cyclic GC is not yet
integrated, we do not manage pointer variables on the stack and closures do
not see their lifetime extended as it should.