Remove useless derivated type from foreign module
This will remove __PTR_ and __ARR*_ symbols from module interfaces
because they can be retrieved directly via T.ptr or T.array.

For function types only occurences that have a matching pointer type are kept.
This is done such that we keep nameable function types that can be used
for closures.