Fetch dynamic type when reading a foreign address
Instead of sticking to the declared type, when reading a foreign address
we now try to get the real allocated datatype and use it as the proxy
type. This can be very useful when a function return a pointer to a base
type, but the data itself has a derivated type.
Also we use direct composite proxy when liballocs tells us that indexing
doesn't make sense (this breaks old code, but it makes more sense in a
dynamically typed environment).

We also had to update the repr cycle breaking policy. Now we are
dereferencing an infinite number of addresses but we never read a
foreign composite with a type already seen above (this have changed the
output of some tests).