Improve the import system
We now will import a foreign library foreign.so using `import elflib.foreign`
(any name after elflib is used as a library name and searched into
elflib.__path__).

It also avoid conflicts with the standard python import system when the .so
file is in the same directory as the .py file.

The module elflib also exports the C basic types such as int, char and float
because they do not belong to any specific library.